Decoupled the server from the kernel
- Added a new interface that just defines the step between the server
and the kernel layer (kernel as a framework engine)
- Refactored internally in order to allow this interface
- Fixed minor tests issues
