+++ src/Adapter/DriftKernel/SyncKernelException.php
@@ -1,0 +1,33 @@
+<?php
+
+/*
+ * This file is part of the Drift Server
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * Feel free to edit as you please, and have fun.
+ *
+ * @author Marc Morera <yuhu@mmoreram.com>
+ */
+
+declare(strict_types=1);
+
+namespace Drift\Server\Adapter\DriftKernel;
+
+use Drift\HttpKernel\AsyncKernel;
+use Drift\Server\Exception\KernelException;
+
+/**
+ * Class SyncKernelException.
+ */
+final class SyncKernelException extends KernelException
+{
+    /**
+     * Build a new instance.
+     */
+    public static function build(): SyncKernelException
+    {
+        return new self(sprintf('Your kernel MUST implement %s', AsyncKernel::class));
+    }
+}
