+++ src/Exception/KernelException.php
@@ -15,19 +15,11 @@ declare(strict_types=1);
 
 namespace Drift\Server\Exception;
 
-use Drift\HttpKernel\AsyncKernel;
 use Exception;
 
 /**
- * Class SyncKernelException.
+ * Class KernelException.
  */
-final class SyncKernelException extends Exception
+class KernelException extends Exception
 {
-    /**
-     * Build a new instance.
-     */
-    public static function build(): SyncKernelException
-    {
-        return new self(sprintf('Your kernel MUST implement %s', AsyncKernel::class));
-    }
 }
