+++ tests/KernelEventsTest.php
@@ -51,7 +51,7 @@ class KernelEventsTest extends TestCase
     public function testShutdown()
     {
         $phpVersion = phpversion();
-        if (strpos($phpVersion, '8.') === 0) {
+        if (0 === strpos($phpVersion, '8.')) {
             $this->markTestSkipped('Events are not tested yet in PHP8.');
         }
 
@@ -82,7 +82,7 @@ class KernelEventsTest extends TestCase
     public function testShutdownIsForced()
     {
         $phpVersion = phpversion();
-        if (strpos($phpVersion, '8.') === 0) {
+        if (0 === strpos($phpVersion, '8.')) {
             $this->markTestSkipped('Events are not tested yet in PHP8.');
         }
